fix: moved interfaces into separate files
