+++ packages/salesforcedx-vscode-core/src/commands/forceSourceRetrieveMetadata/index.ts
@@ -4,34 +4,7 @@
  * Licensed under the BSD 3-Clause license.
  * For full license text, see LICENSE.txt file in the repo root or https://opensource.org/licenses/BSD-3-Clause
  */
-import { LocalComponent } from '@salesforce/salesforcedx-utils-vscode';
 export { RetrieveDescriberFactory } from './describers';
 export { forceSourceRetrieveCmp } from './forceSourceRetrieveCmp';
-
-/**
- * Provides information for force.source.retrieve.component execution
- */
-export interface RetrieveDescriber {
-  /**
-   * Builds the force:source:retrieve metadata argument
-   * @param data optional data to use while building the argument
-   * @returns parameter for metadata argument (-m)
-   */
-  buildMetadataArg(data?: LocalComponent[]): string;
-
-  /**
-   * Gather list of components to be retrieved
-   * @returns local representations of components
-   */
-  gatherOutputLocations(): Promise<LocalComponent[]>;
-}
-
-/**
- * An object capable of triggering the force:source:retrieve metadata command
- */
-export interface RetrieveMetadataTrigger {
-  /**
-   * The RetrieveDescriber to use for the retrieve execution
-   */
-  describer(): RetrieveDescriber;
-}
+export { RetrieveDescriber } from './retrieveDescriber';
+export { RetrieveMetadataTrigger } from './retrieveMetadataTrigger';
